backend/requirements.txt
pandas
numpy
yfinance
matplotlib
seaborn
flask
flask-cors
os_sys
openai
python-dotenv